Some cleanup. Map 'hitting' to 'batting' in scraping targets. Remove re-try code in player stats.

+++ processing/prune_missing_players.py
@@ -9,11 +9,14 @@ with open(os.path.join("missing_targets.json")) as f:
         target = missing_targets[player_id]
         ext_player_id = target['External Player Id']
         path = os.path.join("..", "data", "players", ext_player_id)
+        if not os.path.exists(path):
+            print("{} does not exist. Skipping...".format(path))
+            continue
         print("Removing " + path)
         try:
             shutil.rmtree(path)
         except:
-            print("Couldn't remove {}. Already removed?".format(path))
+            print("Couldn't remove {}.".format(path))
